Collapse text creators into one class
