+++ ui-followed/src/main/java/app/tivi/home/followed/Followed.kt
@@ -55,7 +55,7 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.items
-import app.tivi.common.compose.LocalHomeTextCreator
+import app.tivi.common.compose.LocalTiviTextCreator
 import app.tivi.common.compose.RefreshButton
 import app.tivi.common.compose.SearchTextField
 import app.tivi.common.compose.SortMenuPopup
@@ -176,7 +176,7 @@ private fun FollowedShowItem(
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val textCreator = LocalHomeTextCreator.current
+    val textCreator = LocalTiviTextCreator.current
     Row(
         modifier
             .clickable(onClick = onClick)
