+++ ui-recommended/src/main/java/app/tivi/home/recommended/RecommendedShowsFragment.kt
@@ -27,15 +27,15 @@ import androidx.core.net.toUri
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import app.tivi.common.compose.LocalHomeTextCreator
 import app.tivi.common.compose.LocalTiviDateFormatter
+import app.tivi.common.compose.LocalTiviTextCreator
 import app.tivi.common.compose.paging.collectAsLazyPagingItems
 import app.tivi.common.compose.shouldUseDarkColors
 import app.tivi.common.compose.theme.TiviTheme
 import app.tivi.extensions.DefaultNavOptions
-import app.tivi.home.HomeTextCreator
 import app.tivi.settings.TiviPreferences
 import app.tivi.util.TiviDateFormatter
+import app.tivi.util.TiviTextCreator
 import com.google.accompanist.insets.LocalWindowInsets
 import com.google.accompanist.insets.ViewWindowInsetObserver
 import dagger.hilt.android.AndroidEntryPoint
@@ -44,7 +44,7 @@ import javax.inject.Inject
 @AndroidEntryPoint
 class RecommendedShowsFragment : Fragment() {
     @Inject internal lateinit var tiviDateFormatter: TiviDateFormatter
-    @Inject internal lateinit var homeTextCreator: HomeTextCreator
+    @Inject internal lateinit var tiviTextCreator: TiviTextCreator
     @Inject lateinit var preferences: TiviPreferences
 
     private val viewModel: RecommendedShowsViewModel by viewModels()
@@ -67,7 +67,7 @@ class RecommendedShowsFragment : Fragment() {
         setContent {
             CompositionLocalProvider(
                 LocalTiviDateFormatter provides tiviDateFormatter,
-                LocalHomeTextCreator provides homeTextCreator,
+                LocalTiviTextCreator provides tiviTextCreator,
                 LocalWindowInsets provides windowInsets,
             ) {
                 TiviTheme(useDarkColors = preferences.shouldUseDarkColors()) {
