+++ ui-showdetails/build.gradle
@@ -68,7 +68,6 @@ dependencies {
     implementation Libs.AndroidX.Lifecycle.viewmodel
 
     implementation Libs.AndroidX.coreKtx
-    implementation Libs.AndroidX.emoji
     implementation Libs.AndroidX.Fragment.fragment
     implementation Libs.AndroidX.Fragment.fragmentKtx
     implementation Libs.AndroidX.Navigation.fragment
