+++ ui-showdetails/src/main/java/app/tivi/showdetails/details/ShowDetailsFragment.kt
@@ -31,6 +31,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.navOptions
+import app.tivi.common.compose.LocalTiviTextCreator
 import app.tivi.common.compose.LogCompositions
 import app.tivi.common.compose.shouldUseDarkColors
 import app.tivi.common.compose.theme.TiviTheme
@@ -38,6 +39,7 @@ import app.tivi.extensions.DefaultNavOptions
 import app.tivi.extensions.viewModelProviderFactoryOf
 import app.tivi.settings.TiviPreferences
 import app.tivi.util.TiviDateFormatter
+import app.tivi.util.TiviTextCreator
 import com.google.accompanist.insets.LocalWindowInsets
 import com.google.accompanist.insets.ViewWindowInsetObserver
 import dagger.hilt.android.AndroidEntryPoint
@@ -47,7 +49,7 @@ import javax.inject.Inject
 @AndroidEntryPoint
 class ShowDetailsFragment : Fragment() {
     @Inject internal lateinit var vmFactory: ShowDetailsFragmentViewModel.Factory
-    @Inject internal lateinit var textCreator: ShowDetailsTextCreator
+    @Inject internal lateinit var textCreator: TiviTextCreator
     @Inject internal lateinit var tiviDateFormatter: TiviDateFormatter
     @Inject lateinit var preferences: TiviPreferences
 
@@ -108,7 +110,7 @@ class ShowDetailsFragment : Fragment() {
 
         setContent {
             CompositionLocalProvider(
-                LocalShowDetailsTextCreator provides textCreator,
+                LocalTiviTextCreator provides textCreator,
                 LocalWindowInsets provides windowInsets,
             ) {
                 TiviTheme(useDarkColors = preferences.shouldUseDarkColors()) {
